Add lock around bosh-lite when deploying CF release
We want to ensure that only one cf-release deployment is happening at a
time. We don't want to run into situations where we are deploying
cf-release with the flag ON at the same time that the job to deploy it
with the flag OFF kicks off. The same about the CATs.
So because of that, the whole "cf-release" group runs inside a lock.

This also adds semver resource to cf-release, so we don't need to delete
the deployment and releases to deploy the same number again.

All locks and semver are stored right now in the `metadata` branch of
this repo.