Fix triggering of the bits-service-release pipeline
Using git resources as triggers for intermediate jobs isn't a good idea
because if you trigger the first job manually it won't carry on to the
next one (the resource has already triggered it before).
This changes following jobs to trigger by the semver resource, that we
always modify in the first step.

[#112610587]